+++ agrest-base/src/main/java/io/agrest/base/jsonvalueconverter/DefaultJsonValueConverterFactoryProvider.java
@@ -5,6 +5,7 @@ import org.apache.cayenne.di.DIRuntimeException;
 import org.apache.cayenne.di.Inject;
 import org.apache.cayenne.di.Provider;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
@@ -48,6 +49,7 @@ public class DefaultJsonValueConverterFactoryProvider implements Provider<IJsonV
         converters.put(Float.class, FloatConverter.converter());
         converters.put(float.class, FloatConverter.converter());
         converters.put(Double.class, DoubleConverter.converter());
+        converters.put(BigDecimal.class, BigDecimalConverter.converter());
         converters.put(double.class, DoubleConverter.converter());
         converters.put(Long.class, LongConverter.converter());
         converters.put(long.class, LongConverter.converter());